// frontend/person-blog-app/packages/blog/src/shared/api/generated/models/textPostDetailResponse.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Gateway.API
 * OpenAPI spec version: 1.0
 */
import type { TextPostAuthor } from './textPostAuthor';
import type { TextPostCategory } from './textPostCategory';
import type { TextPostMedia } from './textPostMedia';
import type { TextPostViewerState } from './textPostViewerState';

export interface TextPostDetailResponse {
  id?: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  lead?: string | null;
  /** @nullable */
  text?: string | null;
  createdAt?: string;
  estimatedReadingTimeMinutes?: number;
  viewCount?: number;
  likeCount?: number;
  dislikeCount?: number;
  /** @nullable */
  categories?: TextPostCategory[] | null;
  /** @nullable */
  media?: TextPostMedia[] | null;
  author?: TextPostAuthor;
  viewer?: TextPostViewerState;
}